+++ packages/cli/src/commands/init.ts
@@ -1,7 +1,6 @@
 import { writeFileSync } from "node:fs";
 import { configName } from "@yummacss/nitro";
-import { feedback } from "@/utils/feedback";
-import { cli } from "@/utils/status";
+import { logger } from "@/utils/logger";
 
 function generateConfig(): { filename: string; content: string } {
 	return {
@@ -11,14 +10,14 @@ function generateConfig(): { filename: string; content: string } {
 }
 
 export function init() {
-	const status = cli.progress("Initialize the configuration...");
+	const status = logger.init.start();
 
 	try {
 		const { filename, content } = generateConfig();
 		writeFileSync(filename, content);
-		status.succeed(feedback.init.success);
+		status.succeed(logger.init.success(filename));
 	} catch (_error) {
-		status.fail(feedback.init.fail);
+		status.fail(logger.init.fail());
 		process.exit(1);
 	}
 }